Fix cinder volume type creation and rename variable
Commit 7062b12b9b91a1eb405d7079c62b62b0de59109f introduced
a bug where the default_volume_type is set to 'lvm' but
that is not the volume-type that's actually created, that
is 'lvmdriver-1' in gate runs, so we see a ton of
VolumeTypeNotFoundByName for the 'lvm' volume type in the
cinder-api logs.

This simply sets the default_volume_type to the volume
type that's created in the create_volume_types function.

Also renames the default_type variable to default_name to
be consistent with variable naming in the rest of the
cinder script.

Change-Id: Ib35ffc09f6582354010ac8ca35da581ea508877a
Closes-Bug: #1348786
(cherry picked from commit 6a4aa78fa3728f1b0a98850f432211da5c8caedf)